requirements.txt
certifi==2021.10.8
requests==2.27.1
yarl==1.7.2
colorama==0.4.4
dnspython==2.2.1
asyncstdlib==3.10.0
aiohttp>=3.8.0
aiohttp_socks==0.7.1
async-timeout==4.0.2
uvloop>=0.16.0; sys_platform != "win32" and sys_platform != "cygwin"
pyopenssl==22.0.0
cryptography>=37.0
psutil>=5.9.0
